Move util.js to util/general.js
