+++ app/js/util/general.js
